chore : move error code
